export/html: extract generators to separate files
